Move noop metrics.
