+++ platform-sdk/swirlds-common/src/main/java/com/swirlds/common/metrics/noop/internal/NoOpMetricsFactory.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.swirlds.common.test.metrics.internal;
+package com.swirlds.common.metrics.noop.internal;
 
 import com.swirlds.common.metrics.Counter;
 import com.swirlds.common.metrics.DoubleAccumulator;
