+++ platform-sdk/swirlds-platform-core/src/test/java/com/swirlds/platform/stats/CycleTimingStatTest.java
@@ -20,7 +20,7 @@ import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import com.swirlds.common.metrics.Metrics;
-import com.swirlds.common.test.metrics.NoOpMetrics;
+import com.swirlds.common.metrics.noop.NoOpMetrics;
 import com.swirlds.platform.stats.cycle.CycleDefinition;
 import java.time.temporal.ChronoUnit;
 import java.util.List;
