+++ platform-sdk/swirlds-unit-tests/common/swirlds-common-test/src/test/java/com/swirlds/common/test/NoOpMetricsTest.java
@@ -39,8 +39,8 @@ import com.swirlds.common.metrics.Metrics;
 import com.swirlds.common.metrics.RunningAverageMetric;
 import com.swirlds.common.metrics.SpeedometerMetric;
 import com.swirlds.common.metrics.StatEntry;
+import com.swirlds.common.metrics.noop.NoOpMetrics;
 import com.swirlds.common.statistics.StatsBuffered;
-import com.swirlds.common.test.metrics.NoOpMetrics;
 import java.time.Duration;
 import java.time.temporal.ChronoUnit;
 import java.util.HashSet;
